PR Review changes, changed layouts and split page title function into a helper function
