+++ pages/login/index.js
@@ -1,6 +1,6 @@
 import Image from 'next/image'
 import LoginForm from '../../src/components/LoginForm'
-import LayoutAuth from '../../src/components/layouts/LayoutAuth'
+import LayoutAuth from '../../src/components/layouts/Auth'
 
 const Login = () => {
   return (
